requirements.txt
# Bone Tumor Segmentation Project - Requirements
# Python 3.10+

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Medical Image Processing
monai>=1.3.0
nibabel>=5.0.0

# Scientific Computing
numpy>=1.24.0,<1.25.0  # Compatible with scipy 1.8.1
scipy>=1.8.0,<1.9.0

# Data Processing
pandas>=2.0.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
tensorboardX>=2.6.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0.0

# Optional: For better performance
# ninja  # For faster compilation
# einops  # For tensor operations

# Development Tools (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
